Aplico el parche 97537d5
Solución al issue carm-es#16
Aplicamos el parche carm-es@97537d5 sobre los fuentes originales.

+++ fuentes/inside-services/src/main/java/es/mpt/dsic/inside/service/temporalData/impl/TemporalDataBusinessServiceImpl.java
@@ -147,6 +147,10 @@ public class TemporalDataBusinessServiceImpl implements TemporalDataBusinessServ
   }
 
   private void deleteDir(File folder) {
+    // CARM ### v2.0.8.2
+    if (!folder.isDirectory())
+      return;
+    // CARM 2.0.8.2 ###
     for (final File fileEntry : folder.listFiles()) {
       fileEntry.delete();
     }
